feat : Card Sharing Animation 뷰 구현
